This commit should have been split into three separate ones, but it is too late to change at this point.
First, we decentralized the base64 encoding of the skin data away from the DataProvider and fixed a bug, where the skin data of skull entities wasn't even encoded.
Second, we changed the constructor of the SkullEntity class to be a bit more usable. You no longer need to provide a skin yourself, just a location and the basic data of the skull. The rest will be done in the classes constructor.
And last, we made skull items and skull entities not only saving the player's UUID but also the name. With that, we also don't misuse the skin's ID for storing the player's UUID any longer, but store them in attributes in the SkullEntity class.